+++ tests/test_poetry_dynamic_versioning.py
@@ -1,8 +1,17 @@
+import os
 from pathlib import Path
+
+import pytest
+from dunamai import Version
 
 import poetry_dynamic_versioning as plugin
 
 root = Path(__file__).parents[1]
+
+
+@pytest.fixture
+def config():
+    return plugin._default_config()["tool"]["poetry-dynamic-versioning"]
 
 
 def test__deep_merge_dicts():
@@ -35,3 +44,49 @@ def test__get_config__with_plugin_customizations():
     assert config["vcs"] == "git"
     assert config["style"] == "semver"
     assert config["subversion"]["tag-dir"] == "alt/tags"
+
+
+def test__get_version__defaults(config):
+    v, s = plugin._get_version(config)
+    assert v == Version.from_git()
+    assert s == Version.from_git().serialize()
+
+
+def test__get_version__invalid_vcs(config):
+    config["vcs"] = "invalid"
+    with pytest.raises(ValueError):
+        plugin._get_version(config)
+
+
+def test__get_version__invalid_style(config):
+    config["style"] = "invalid"
+    with pytest.raises(ValueError):
+        plugin._get_version(config)
+
+
+def test__get_version__format_jinja(config):
+    os.environ["FOO"] = "foo"
+    config["format-jinja"] = "{% if true %}v1+{{ env['FOO'] }}{% endif %}"
+    _, v = plugin._get_version(config)
+    assert v == "v1+foo"
+
+
+def test__get_version__format_jinja_with_enforced_style(config):
+    config["format-jinja"] = "{% if true %}1+jinja{% endif %}"
+    config["style"] = "pvp"
+    with pytest.raises(ValueError):
+        plugin._get_version(config)
+
+
+def test__get_version__format_jinja_imports_with_module_only(config):
+    config["format-jinja"] = "{{ math.pow(2, 2) }}"
+    config["format-jinja-imports"] = [{"module": "math"}]
+    _, v = plugin._get_version(config)
+    assert v == "4.0"
+
+
+def test__get_version__format_jinja_imports_with_module_and_item(config):
+    config["format-jinja"] = "{{ pow(2, 3) }}"
+    config["format-jinja-imports"] = [{"module": "math", "item": "pow"}]
+    _, v = plugin._get_version(config)
+    assert v == "8.0"