+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Added the option `format-jinja-imports`.
+
 ## v0.7.0 (2020-05-14)
 
 * Added a CLI mode.